requirements.txt
argon2-cffi==20.1.0
asn1crypto==1.4.0
async-generator==1.10
attrs==20.2.0
autopep8==1.5.4
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
backcall==0.2.0
bleach==3.2.1
boto3==1.14.63
botocore==1.17.63
category-encoders==2.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
idna==2.9
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
msrest==0.6.19
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
ntlm-auth==1.5.0
numpy==1.19.2
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.8
Pygments==2.7.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.23.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
SharePlum==0.5.1
six==1.15.0
snowflake-connector-python==2.3.3
snowflake-sqlalchemy==1.2.3
SQLAlchemy==1.3.18
statsmodels==0.12.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1